% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CSIcalc.R
\name{CSIcalc}
\alias{CSIcalc}
\title{Calculate CSI values from salinity object}
\usage{
CSIcalc(sal, scale = 24)
}
\arguments{
\item{sal}{data.frame with Year and Month timestamp columns, with columns of site salinity values. Generally created by one of the CSIimport_ family functions.}

\item{scale}{integer representing the scale at which the CSI will be computed. For example, a value of six would imply that data from the current month and of the past five months will be used for computing the CSI value for a given month.}
}
\value{
A 3D array of CSI values with dimensions of number of months covered, scale of months analysed (typically 1-24), and number of sites.
}
\description{
Calculate monthly CSI values from salinity object for a set of sites over a range of scales.
}
\examples{
# Data file with Year and Month columns
data_path <- system.file("extdata", "Monthly_Coastal_EDEN.csv", package = "CSI")
sal <- CSIimport_monthly(data_path)
csi <- CSIcalc(sal)

}
