% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CSIdist.R
\name{CSIdist}
\alias{CSIdist}
\title{Create distribution plots of CSI values for sites}
\usage{
CSIdist(csi, dir = paste0(getwd(), "/csi_dist"))
}
\arguments{
\item{csi}{array A 3D array of CSI values with dimensions of number of months covered, scale of months analysed (typically 1-24), and number of sites.}

\item{dir}{character Directory to write output files to.}
}
\description{
For each site (layer) in the csi object, plot the intervals (columns) on a distribution plot.
}
\examples{
# Data file with Year and Month columns
data_path <- system.file("extdata", "Monthly_Coastal_EDEN.csv", package = "CSI")
sal <- CSIimport_monthly(data_path)
csi <- CSIcalc(sal)
CSIdist(csi)

}
