% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CSIduration_hydrograph.R
\name{CSIduration_hydrograph}
\alias{CSIduration_hydrograph}
\title{Create salinity duration hydrograph plots for sites}
\usage{
CSIduration_hydrograph(sal, dir = paste0(getwd(),
  "/duration_hydrographs"), thumbs = F, leg = T)
}
\arguments{
\item{sal}{data.frame salinity object with daily or finer resolution salinity data stored in daily_data attribute. Created by CSIimport_daily, CSIimport_unit, or CSIimport_NERRS functions.}

\item{dir}{character Directory to write output files to.}

\item{thumbs}{logical. If true, thumbnail plots will be generated in addition to full-sized plots.}

\item{leg}{logical. If true, legend will be displayed in upper left corner.}
}
\description{
For each site in the salinity object, plot the salinity duration hydrograph.
}
\examples{
# Data file with Year, Month, and Day columns
data_path <- system.file("extdata", "Daily_Waccamaw_LittleBlack_Rivers.csv", package="CSI")
sal <- CSIimport_daily(data_path)
CSIduration_hydrograph(sal)

}
