% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CSIimport.R
\name{CSIimport}
\alias{CSIimport}
\title{Import files for CSI calculation}
\usage{
CSIimport(file, interval = c("daily", "hourly", "15-minute", "60-minute",
  "irregular", "interval"), scale = 24)
}
\arguments{
\item{file}{character The file to import.}

\item{interval}{character The timestamp frequency -- '15 minute', '60 minute', or 'daily' (default).}

\item{scale}{numeric The range of months over which to compare. Default: 24.}
}
\value{
An array of CSI values
}
\description{
Read in salinity values file.
}
\examples{
csi <- CSIimport("~/Desktop/R/CSI_source/Monthly_Waccamaw_LittleBlack_Rivers.csv")

}
