% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CSIimport_hourly.R
\name{CSIimport_60minute}
\alias{CSIimport_60minute}
\title{Import hourly-value files for CSI calculation}
\usage{
CSIimport_60minute(file)
}
\description{
Read in hourly salinity values file.

Alias of CSIimport_hourly.
}
