% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CSIimport_daily.R
\name{CSIimport_daily}
\alias{CSIimport_daily}
\title{Import daily-value files for CSI calculation}
\usage{
CSIimport_daily(file)
}
\arguments{
\item{file}{character The daily-value file to import. Must have Year, Month, and Day columns (or single column 'Date' = YYYY-MM-DD), and columns of salinty values by site.}
}
\value{
A data.frame of year+monthly timestamps with columns of salinity values
}
\description{
Read in daily salinity values file.
}
\examples{
#add file with Year,Month,Day columns
csi <- CSIimport_daily(system.file("extdata","Daily_Waccamaw_LittleBlack_Rivers.csv",package="CSI"))

}
