% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CSIimport_hourly.R
\name{CSIimport_hourly}
\alias{CSIimport_hourly}
\title{Import hourly-value files for CSI calculation}
\usage{
CSIimport_hourly(file)
}
\arguments{
\item{file}{character The hourly-value file to import. Must have Year, Month, Day, and Hour columns, and columns of salinty values by site.}
}
\value{
A data.frame of year+monthly timestamps with columns of salinity values
}
\description{
Read in hourly salinity values file.
}
\examples{
#csi <- CSIimport_hourly("") #add file with Year,Month,Day,Hour columns

}
