% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CSIimport_monthly.R
\name{CSIimport_monthly}
\alias{CSIimport_monthly}
\title{Import monthly-value files for CSI calculation}
\usage{
CSIimport_monthly(file)
}
\arguments{
\item{file}{character The monthly-value file to import. Must have Year and Month columns, or single YearMo column = 'YYYY-MM', and columns of salinty values by site.}
}
\value{
A salinity object data.frame for calculating CSI values; has Year and Month timestamp columns, with (optionally multiple) individual columns of site salinity values.
}
\description{
Read in monthly salinity values file.
}
\examples{
# Data file with Year and Month columns
data_path <- system.file("extdata", "Monthly_Coastal_EDEN.csv", package = "CSI")
sal <- CSIimport_monthly(data_path)

# Data file with single YearMo column
data_path <- system.file("extdata", "Monthly_ACE_Basin.csv", package = "CSI")
sal <- CSIimport_monthly(data_path)

}
