% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CSIspec_con.R
\name{CSIspec_con}
\alias{CSIspec_con}
\title{Convert specific conductance to salinity}
\usage{
CSIspec_con(sc)
}
\arguments{
\item{sc}{data.frame with Year and Month timestamp columns (e.g., produced by CSIimport_ family functions), with columns of site specific conductance values.}
}
\value{
A data.frame with Year and Month timestamp columns, with columns of site salinity values.
}
\description{
Convert specific conductance values to salinity values for a set of sites.
}
\examples{
# Data file with single Date column, and specific conductance values
data_path <- system.file("extdata", "Daily_South_Carolina_spec_con.csv", package="CSI")
sc <- CSIimport_daily(data_path)
sal <- CSIspec_con(sc)

}
