% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CSIstack.R
\name{CSIstack}
\alias{CSIstack}
\title{Create stacked CSI interval plots for sites}
\usage{
CSIstack(csi, dir = paste0(getwd(), "/csi_stacked"), thumbs = F,
  grouped = T, leg = T)
}
\arguments{
\item{csi}{array A 3D array of CSI values with dimensions of number of months covered, scale of months analysed (typically 1-24), and number of sites.}

\item{dir}{character Directory to write output files to.}

\item{thumbs}{logical. If true, thumbnail plots will be generated in addition to full-sized plots.}

\item{grouped}{logical. If true, second y-asix (salinity) will be the same across all output plots.}

\item{leg}{logical. If true, legend will be displayed in upper left corner.}
}
\description{
For each site (layer) in the csi object, plot the intervals (columns) on a stacked plot.
}
\examples{
# Data file with Year and Month columns
data_path <- system.file("extdata", "Monthly_Coastal_EDEN.csv", package = "CSI")
sal <- CSIimport_monthly(data_path)
csi <- CSIcalc(sal)
CSIstack(csi)

}
