% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CSIupdown.R
\name{CSIupdown}
\alias{CSIupdown}
\title{Create individual CSI interval up/down plots for sites}
\usage{
CSIupdown(csi, dir = paste0(getwd(), "/csi_updowns"))
}
\arguments{
\item{csi}{array A 3D array of CSI values with dimensions of number of months covered, scale of months analysed (typically 1-24), and number of sites.}

\item{dir}{character Directory to write output files to.}
}
\description{
For each site (layer) in the csi object, plot each interval (column).
}
\examples{
# Data file with Year and Month columns
data_path <- system.file("extdata", "Monthly_Coastal_EDEN.csv", package = "CSI")
sal <- CSIimport_monthly(data_path)
csi <- CSIcalc(sal)
CSIupdown(csi)

}
