% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CSIwrite.R
\name{CSIwrite}
\alias{CSIwrite}
\title{Write monthly CSI values from 3D csi object to files}
\usage{
CSIwrite(csi, dir = paste0(getwd(), "/csi_values"))
}
\arguments{
\item{csi}{array A 3D array of CSI values with dimensions of number of months covered, scale of months analysed (typically 1-24), and number of sites.}

\item{dir}{character Directory to write output files to.}
}
\description{
Write monthly CSI values from 3D csi object to text files, one file per site.
}
\examples{
# Data file with Year and Month columns
data_path <- system.file("extdata", "Monthly_Coastal_EDEN.csv", package = "CSI")
sal <- CSIimport_monthly(data_path)
csi <- CSIcalc(sal)
CSIwrite(csi)

}
