% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CSIimport.R
\name{CSIimport}
\alias{CSIimport}
\title{Import salinity value files.}
\usage{
CSIimport(file, interval = c("monthly", "daily", "24hour", "interval",
  "unit"))
}
\arguments{
\item{file}{character The file to import.}

\item{interval}{character The timestamp frequency -- 'daily', 'interval'/'unit', or 'monthly' (default).}
}
\value{
A data.frame with Year and Month timestamp columns, with columns of site salinity values.
}
\description{
Read in salinity values files of various intervals.
}
\examples{
# Data file with Year and Month columns
data_path <- system.file("extdata", "Monthly_Coastal_EDEN.csv", package = "CSI")
sal <- CSIimport(data_path, interval = "monthly")

}
