% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CSIimport.R
\name{CSIimport}
\alias{CSIimport}
\title{Import salinity value files.}
\usage{
CSIimport(file, interval = c("monthly", "daily", "hourly", "15-minute",
  "60-minute", "irregular", "interval"))
}
\arguments{
\item{file}{character The file to import.}

\item{interval}{character The timestamp frequency -- '15 minute', '60 minute', 'daily', 'interval', or 'monthly' (default).}
}
\value{
A data.frame of year+monthly timestamps with columns of salinity values
}
\description{
Read in salinity values files of various intervals.
}
\examples{
csi <- CSIimport("~/Desktop/R/CSI_source/Monthly_Waccamaw_LittleBlack_Rivers.csv")

}
