% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CSIimport_daily.R
\name{CSIimport_daily}
\alias{CSIimport_daily}
\title{Import daily-value files for CSI calculation}
\usage{
CSIimport_daily(file, mindays = 15)
}
\arguments{
\item{file}{character The daily-value file to import. Must have Year, Month, and Day columns, or single column Date = 'YYYY-MM-DD', and columns of salinty values by site.}

\item{mindays}{integer the minimum number of daily values needed to produce a mean value for a given month (default 15). Months with fewer data values than this will be asigned mean NA.}
}
\value{
A salinity object data.frame for calculating CSI values; has Year and Month timestamp columns, with (optionally multiple) individual columns of site salinity values.
}
\description{
Read in daily salinity values file.
}
\examples{
# Data file with Year, Month, and Day columns
data_path <- system.file("extdata", "Daily_Waccamaw_LittleBlack_Rivers.csv", package="CSI")
sal <- CSIimport_daily(data_path)

# Data file with single Date column
data_path <- system.file("extdata", "Daily_South_Carolina.csv", package="CSI")
sal <- CSIimport_daily(data_path)

}
