% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CSIimport_unit.R
\name{CSIimport_interval}
\alias{CSIimport_interval}
\title{Import unit-value files for CSI calculation}
\usage{
CSIimport_interval(file)
}
\arguments{
\item{file}{character The unit-value file to import. Must have Year, Month, Day, and Time columns, or single Timestamp = 'YYYY-MM-DD HH:mm:ss', and columns of salinty values by site.}
}
\description{
Read in unit value salinity file.

Alias of CSIimport_unit.
}
