% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CSIimport_monthly.R
\name{CSIimport_monthly}
\alias{CSIimport_monthly}
\title{Import monthly-value files for CSI calculation}
\usage{
CSIimport_monthly(file)
}
\arguments{
\item{file}{character The monthly-value file to import. Must have Year and Month columns, and columns of salinty values by site.}
}
\value{
A data.frame of year+monthly timestamps with columns of salinity values
}
\description{
Read in monthly salinity values file.
}
\examples{
csi <- CSIimport_monthly("~/Desktop/R/CSI_source/Monthly_Waccamaw_LittleBlack_Rivers.csv")

}
