% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CSIimport_unit.R
\name{CSIimport_unit}
\alias{CSIimport_unit}
\title{Import unit-value files for CSI calculation}
\usage{
CSIimport_unit(file)
}
\arguments{
\item{file}{character The unit-value file to import. Must have Year, Month, Day, and Time columns, or single Timestamp = 'YYYY-MM-DD HH:mm:ss', and columns of salinty values by site.}
}
\value{
A data.frame with Year and Month timestamp columns, with columns of site salinity values.
}
\description{
Read in unit value salinity file.
}
\examples{
#ADD
# Data file with Year, Month, Day, and Hour columns
# data_path <- system.file("extdata", "", package="CSI")
# sal <- CSIimport_unit(data_path)

# Data file with single Timestamp column
data_path <- system.file("extdata", "Unitvalue_North_Inlet.csv", package="CSI")
sal <- CSIimport_unit(data_path)

}
